Add scripts for terminating processes on port 4567 and SSE vs Polling comparison
This commit introduces a new script, `kill_port_4567.sh`, to facilitate
the termination of any processes using port 4567. This is useful for
ensuring that port is free for new services or applications to start.
Additionally, a markdown document comparing Server-Sent Events (SSE)
with Polling for the City Services Redis Message Monitor is added.
This comparison helps to clarify the advantages and disadvantages of
both methods, guiding developers in making informed choices for real-time
monitoring.

No issue references were provided for this commit.
